correct the trust engine dashboard route and move the files to the correct location
