--- src/app/(dashboard)/trust-engine/page.tsx
@@ -1,12 +1,0 @@
-import TrustEngineScreen from "@/src/components/Dashboard/TrustEngine/TrustEngineScreen"
-import React from "react"
-
-const TrustEnginePage = () => {
-  return (
-    <div>
-      <TrustEngineScreen />
-    </div>
-  )
-}
-
-export default TrustEnginePage
