+++ src/app/(home)/trust-engine/dashboard/page.tsx
@@ -1,0 +1,12 @@
+import TrustEngineScreen from "@/src/components/TrustEngine/TrustEngineScreen"
+import React from "react"
+
+const TrustEngineDashboard = () => {
+  return (
+    <div>
+      <TrustEngineScreen />
+    </div>
+  )
+}
+
+export default TrustEngineDashboard
